requirements.txt
streamlit>=1.28.0
openai>=1.0.0
streamlit-mic-recorder>=0.0.8
pydub>=0.25.1
python-dotenv>=1.0.0
